chore: move liveVersionedHome to rollback.go
